Pass user settings to playbook
OpenStack Ansible relies on some user settings to be defined, such as
user secrets for instance, so we need to read them and pass to playbook.
Otherwise, the playbooks will fail due to undefined variables.

The settings to be read are hardcoded the same way they are hardcoded
in bootstrap-ansible.sh [1].

[1] http://git.openstack.org/cgit/openstack/openstack-ansible/tree/scripts/bootstrap-ansible.sh?h=stable/newton#n183