Stop chain execution on error
If by some reason a playbook has finished with errors we have to
immediately stop further execution and allow operator to fix issues.
In order to do that we need to raise an exception if exit code of
playbook execution is not 0, because it's treated by Celery as
failed execution.