Implement alternative openstack-ansible driver
The reason behind alternative implementation is that Ansible is too buggy
as a library. For instance, there are a lot of resource leaks such as
open pipes and so on. For instance, when we create many PlaybookExecutor
instances, leaked pipes cause 'Too many open files' runtime error. It
could be probably avoided by increasing ulimits, though I'm not sure
it's a good thing to maintain a lot of hacks.

The major difference of alternative implementation is that it runs Ansible
playbooks using shell execution rather than programmatic one. Hence,
when program is finished all leaked pipes are free by OS.